requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
celery-with-mongodb==3.0
dj-database-url==0.4.2
Django==1.4.8
django-appconf==1.0.2
django-celery==3.2.2
django-compressor==1.4
-e git://github.com/dcramer/django-devserver@35f9abf601e6cc2ed3cdbb78fb60018d7bd5a48d#egg=django_devserver
django-extensions==1.4.5
django-mptt==0.7.4
django-recurrence==1.0.3
django-social-auth==0.7.28
django-statsd-mozilla==0.4.0
django-tastypie==0.9.9
django-tinymce==1.5.1a1
factory-boy==2.9.2
Faker==0.8.7
gunicorn==19.7.1
guppy==0.1.10
httplib2==0.10.3
ipaddress==1.0.18
kombu==3.0.37
mimeparse==0.1.3
oauth2==1.9.0.post1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pymongo==3.5.1
python-dateutil==1.5
python-digest==1.7
python-mimeparse==1.6.0
python-openid==2.2.5
pytz==2017.3
rcssmin==1.0.6
rjsmin==1.0.12
six==1.4.0
South==0.7.6
sqlparse==0.2.4
statsd==3.2.1
text-unidecode==1.1
typing==3.6.2
urllib3==1.22
Werkzeug==0.12.2
